Tests: Build and test KWIML even when testing an external CMake
The CMake_TEST_EXTERNAL_CMAKE option added by commit 9608ef6f (Tests:
Optionally configure tests exclusively, 2014-03-03) is intended to allow
one to run the CMake test suite with a compiler that may not be
supported for hosting the build of CMake itself.  However, we currently
use the CMake test infrastructure to test KWIML everywhere that CMake
supports.  In order to continue testing KWIML even in places that CMake
itself does not build, include it even when testing an external CMake.