cmcurl: Fix a build failure with the Cray compiler on Linux (#15026)
The error is caused by a workaround for UNICOS.  The workaround
incorrectly uses the _CRAYC macro to determine if building for
UNICOS.  This macro will always be defined for the Cray C and
C++ compiler regardless of the target platform, even when
building for Linux.  The correct macro should be _CRAY, which as
per Cray documentation is only defined for UNICOS targeted
builds.